@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "org.ajcm.hiad"
         minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 24
-        versionName "1.9.4"
+        targetSdkVersion 28
+        versionCode 28
+        versionName "1.9.30"
     }
     buildTypes {
         release {
@@ -18,29 +18,29 @@ android {
 }
 
 dependencies {
-    def VERSION_ANDROID_SUPPORT = '27.1.1'
-    def VERSION_FIREBASE = '12.0.1'
+    def VERSION_ANDROID_SUPPORT = '28.0.0'
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "com.android.support:appcompat-v7:$VERSION_ANDROID_SUPPORT"
-    implementation "com.android.support:support-v4:$VERSION_ANDROID_SUPPORT"
-    implementation "com.android.support:design:$VERSION_ANDROID_SUPPORT"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.0.0'
 
-    implementation "com.google.firebase:firebase-ads:$VERSION_FIREBASE"
-    implementation "com.google.firebase:firebase-analytics:$VERSION_FIREBASE"
-    implementation "com.google.firebase:firebase-storage:$VERSION_FIREBASE"
-    implementation "com.google.firebase:firebase-auth:$VERSION_FIREBASE"
-    implementation "com.google.firebase:firebase-messaging:$VERSION_FIREBASE"
+    implementation "com.google.firebase:firebase-ads:17.1.3"
+    implementation "com.google.firebase:firebase-analytics:16.3.0"
+    implementation "com.google.firebase:firebase-storage:16.1.0"
+    implementation "com.google.firebase:firebase-auth:16.1.0"
+    implementation "com.google.firebase:firebase-messaging:17.4.0"
 
     implementation 'com.sothree.slidinguppanel:library:3.2.0'
     implementation 'fr.avianey.com.viewpagerindicator:library:2.4.1.1@aar'
     implementation 'org.zakariya.stickyheaders:stickyheaders:0.7.6@aar'
     implementation 'com.github.ivbaranov:materialfavoritebutton:0.1.4'
     implementation 'com.github.chrisbanes:PhotoView:2.0.0'
-    implementation 'com.github.bumptech.glide:glide:4.8.0'
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
 
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'com.github.bumptech.glide:glide:4.10.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.10.0'
+
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
 }
 
