@@ -1,20 +1,16 @@
 package org.ajcm.hiad.activities;
 
-import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.design.widget.TabLayout;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import com.google.android.material.tabs.TabLayout;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.MenuItem;
 import android.view.Window;
 import android.view.WindowManager;
 
 import org.ajcm.hiad.R;
-import org.ajcm.hiad.adapters.MusicViewAdapter;
 import org.ajcm.hiad.adapters.PagerAdapter;
 import org.ajcm.hiad.fragments.MusicFragment;
 
