@@ -1,7 +1,7 @@
 package org.ajcm.hiad.activities;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import org.ajcm.hiad.R;
