@@ -1,13 +1,10 @@
 package org.ajcm.hiad.adapters;
 
-import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import org.ajcm.hiad.views.WelcomeFragment;
-
-import java.util.ArrayList;
 
 /**
  * Created by jhonlimaster on 08-12-16.
