@@ -2,7 +2,7 @@ package org.ajcm.hiad.views;
 
 import android.annotation.TargetApi;
 import android.os.Build;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.view.View;
 
 import org.ajcm.hiad.R;
