@@ -1,19 +1,18 @@
 package org.ajcm.hiad.activities;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import androidx.annotation.NonNull;
+import androidx.core.widget.ContentLoadingProgressBar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
+import android.view.View;
+import android.widget.Toast;
 
-import com.bumptech.glide.Glide;
-import com.bumptech.glide.annotation.GlideModule;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
-import com.bumptech.glide.module.AppGlideModule;
 import com.github.chrisbanes.photoview.PhotoView;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
@@ -21,11 +20,13 @@ import com.google.firebase.storage.StorageReference;
 import org.ajcm.hiad.R;
 import org.ajcm.hiad.utils.GlideApp;
 
-public class PartitureActivity extends AppCompatActivity implements OnCompleteListener {
+public class PartitureActivity extends AppCompatActivity {
 
     private static final String TAG = "PartitureActivity";
 
     private PhotoView photoView;
+    private StorageReference himnoRef;
+    private ContentLoadingProgressBar progressBar;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -36,27 +37,45 @@ public class PartitureActivity extends AppCompatActivity implements OnCompleteLi
 
         String urlFirebase = getResources().getString(R.string.url_firebase);
         FirebaseStorage storage = FirebaseStorage.getInstance();
+        storage.setMaxOperationRetryTimeMillis(5000);
         StorageReference reference = storage.getReferenceFromUrl(urlFirebase);
         int numeroHimno = getIntent().getExtras().getInt(MainActivity.NUMERO);
 
-        StorageReference himnoRef = reference.child("partituras/himno" + formatNumber(numeroHimno) + ".jpg");
+        himnoRef = reference.child("partituras/himno" + formatNumber(numeroHimno) + ".jpg");
 
-        himnoRef.getDownloadUrl().addOnCompleteListener(this).addOnFailureListener(new OnFailureListener() {
+        himnoRef.getDownloadUrl().addOnCompleteListener(new OnCompleteListener<Uri>() {
+            @Override
+            public void onComplete(@NonNull Task<Uri> task) {
+                progressBar.hide();
+                photoView.setVisibility(View.VISIBLE);
+
+                // hide status bar
+                View decorView = getWindow().getDecorView();
+                int uiOptions = View.SYSTEM_UI_FLAG_FULLSCREEN;
+                decorView.setSystemUiVisibility(uiOptions);
+
+                if (task.isSuccessful()) {
+                    GlideApp.with(getApplicationContext())
+                            .load(task.getResult().toString())
+                            .diskCacheStrategy(DiskCacheStrategy.ALL)
+                            .into(photoView);
+                } else {
+                    Toast.makeText(PartitureActivity.this, "Upps, vuelva a intentarlo.", Toast.LENGTH_SHORT).show();
+                    finish();
+                }
+            }
+        }).addOnFailureListener(new OnFailureListener() {
             @Override
             public void onFailure(@NonNull Exception e) {
                 finish();
+                Log.e(TAG, "onFailure: ");
             }
         });
+        progressBar = findViewById(R.id.loading_progress_bar);
+        photoView.setVisibility(View.GONE);
+        progressBar.show();
 
 //        StorageReference himnoRef = reference.child("partituras/himno" + number + ".ogg");
-    }
-
-    @Override
-    public void onComplete(@NonNull Task task) {
-        GlideApp.with(this)
-                .load(task.getResult().toString())
-                .diskCacheStrategy(DiskCacheStrategy.ALL)
-                .into(photoView);
     }
 
     private String formatNumber(int numero) {
